Refactor so both modules can be used together
