+++ lib/phlexible/rails/action_controller/render_action.rb
@@ -1,27 +1,19 @@
+# frozen_string_literal: true
+
+
+require_relative "base"
+
 module Phlexible
   module Rails
     module ActionController
       module RenderAction
+        include Base
 
         def render_to_body(options = {})
           if view = phlex_view(options)
             render view.new(**options.except(:action, :prefixes, :template, :layout))
           else
             super
-          end
-        end
-
-        private
-
-        def phlex_view(options)
-          phlex_view_path(options)&.camelize&.safe_constantize
-        end
-
-        def phlex_view_path(options)
-          if options[:action]
-            "/#{controller_path}/#{options[:action]}_view"
-          elsif options[:template]
-            "/#{options[:template]}_view"
           end
         end
       end
